Bug 1374730 - Update webrender bindings for API changes in WR cset 5066a13. r=mrobinson
This change preserves the existing semantics of the webrender_bindings code,
although it may result in creating unnecessary clips. A follow-up bug will be
filed to make clip optional in the C++ interface, and avoid passing it in places
where it is not necessary.

MozReview-Commit-ID: KMeBumpgDXL

--HG--
extra : rebase_source : 9ec04e1e8f0b985532908460958f30ed42e220bd