Bug 1417519 - Hook up APZ to use the hit-testing API. r=botond,mstange
This adds the code that APZ needs to use the WR hit-testing API. For now
(if the pref is enabled) it just does the hit-test and compares the
result to the regular hit-test, printing out discrepancies. It also
doesn't yet get the scrollbar result.

MozReview-Commit-ID: 3pjYSWDGeDr

--HG--
extra : rebase_source : 9766fe3e77dd85f130ce155041cec7c0daae67ec