Bug 1562892 - WR capture native fonts as raw r=lsalzman
we save the native fonts by their full path now. On macOS, there is no
such thing as a full filesystem path for a CGFont (or at least we don't track it),
so loading a capture falls back to the old logic of using the dummy font.

Differential Revision: https://phabricator.services.mozilla.com/D36604

--HG--
extra : moz-landing-system : lando