Bug 1591436 - Avoid using glTexImage3D on android emulator. r=gw
A bug in the android emulator means that glTexImage3D thinks the maximum texture
size is the value of GL_MAX_TEXTURE_SIZE, rather than the value returned by
glGetIntegerv(GL_MAX_TEXTURE_SIZE). This means that calls to glTexImage3D fail
when width, height, or depth > 3379.

This forces us to use glTexStorage3D instead of glTexImage3D, avoiding
this bug. This relies upon swizzling, which is indeed supported on the
android emulator.

Differential Revision: https://phabricator.services.mozilla.com/D51293

--HG--
extra : moz-landing-system : lando