Bug 1589666 - Disable picture caching whilst pinch-zooming. r=gw
Whilst pinch zooming, every picture cache tile gets completely
invalidated every frame. It is therefore a waste of memory bandwidth
to render in to picture cache tiles then composite those to the
screen. This change dynamically disables picture caching for frames in
which we are pinch zooming. The exception is if we are using a native
compositor, in which case picture caching will remain enabled, because
it relies on picture caching to work, and does not waste memory
bandwidth.

Differential Revision: https://phabricator.services.mozilla.com/D52017

--HG--
extra : moz-landing-system : lando