Bug 1594364 - Support subpx AA with picture caching in more cases. r=kvark
Instead of creating a picture caching slice for any content that
is fixed position, also check if the clip(s) for the cluster are
fixed position or anchored to the scroll root.

This prevents WR creating slices for parallax style effects. There's
not much point in doing this anyway, since those slices will invalidate
due to the spatial node of the clip changing during scrolling. This
also allows subpixel AA to be enabled in more situations.

Differential Revision: https://phabricator.services.mozilla.com/D52621

--HG--
extra : moz-landing-system : lando