Bug 1555376 - Snap glyphs before clipping in the shader. r=lsalzman
Snapping glyph positions are an internal detail to a primitive. As such,
any snapping required must be taken into account when calculating the
local rect. That ensures that when the clip is applied, it doesn't cut
off parts of the glyph that would have been retained after snapping.

Differential Revision: https://phabricator.services.mozilla.com/D55348

--HG--
extra : moz-landing-system : lando