Bug 1600593 - Remove snap offset references from the shaders. r=kvark
We now assume that most snapping is done during scene building. This
means that nearly every point and rect used in the shaders are
pre-snapped and nothing more need be done. The text shader is an
exception, but it also effectively pre-snaps because when it calculates
its local rect for a glyph, it takes into account any snapping
requirements, rather than adjusting for snapping after its local rect
interacts with other elements.

As such, we can remove any need for generic snap offsets from the
shaders.

Differential Revision: https://phabricator.services.mozilla.com/D55436

--HG--
extra : moz-landing-system : lando