Bug 1595768 - Keep empty items in the texture cache. r=gw
Adds a notion of empty cache items in the texture cache, that are not uploaded into textures but have a cache entry and expire like other types of entries. The motivation for this is to avoid continuously requesting invalid glyphs to be re-rasterized. Currently if a page contains invalid glyphs we gracefully fail to reasterize it but since we don't keep a trace of it in the cache it appears new each frame which cause us to schedule work on the rayon thread pool every frame at great costs.

Differential Revision: https://phabricator.services.mozilla.com/D56958

--HG--
extra : moz-landing-system : lando