Bug 1608280 - Part 1 - Add a valid_rect to picture cache and composite tiles. r=mstange
This patch introduces a per-tile valid rect. In the initial implementation,
this only uses the bounds of the overall picture cache bounding rect. The
next part of this patch will make use of true per-tile valid regions, to
improve performance where there are holes in a single cache slice.

Differential Revision: https://phabricator.services.mozilla.com/D61378

--HG--
extra : moz-landing-system : lando