Bug 1579235 - Part 1 - Support prefer compositor flag in wrench. r=nical
Add support to the yaml reader and writer to be able to specify
that a primitive should set the PREFER_COMPOSITOR_SURFACE flag.

This flag is not currently used, but in future will signal the
picture caching code to promote a primitive to draw on a native
compositor surface where possible.

Differential Revision: https://phabricator.services.mozilla.com/D62693

--HG--
extra : moz-landing-system : lando