Bug 1529260 - Match shader and CPU glyph rasterization logic for glyphs in local space. r=lsalzman
This patch makes the CPU side incorporate the raster scale when
calculating the subpixel position of a glyph. It also makes the shader
side not include the glyph scale factor when recalculating the glyph
position (since it was not known/used when determining the subpixel
position in the first place). This makes the subpixel position stable
when we transition between Screen and Local(raster_scale) spaces.

Differential Revision: https://phabricator.services.mozilla.com/D62812

--HG--
extra : moz-landing-system : lando