Bug 1616594 - Use the scene's frame builder config during frame building. r=gw
The render backend's frame builder config is kept only in order to send updates to the scene builder's frame builder config which will update the scene's in the next transaction. If need be the scene configs can be updated right away by looping over the documents. This avoids confusing bugs where only updating the backend's config affects the visibility pass but not the rest.

Differential Revision: https://phabricator.services.mozilla.com/D63337

--HG--
extra : moz-landing-system : lando