Bug 1626827 - Force WR picture tasks to fit into max target size r=Bert
this is an attempt to handle tasks outside of the device bounds,
that belong to surfaces not establishing raster roots.

I suspect that the scaling we are now setting up in adjust_scale_for_max_surface_size
doesn't work properly, since the function was assumed to only affect the raster-rooted
surfaces. But it does fix the crash we have.

Differential Revision: https://phabricator.services.mozilla.com/D69654

--HG--
extra : moz-landing-system : lando