Bug 1628564 - Remove the common primitive template data for picture primitives. r=nical
Picture primitives are special, so it doesn't make sense to have
the normal common primitive data for them. For example, the
bounding rect of a picture is determined during frame building.

Removing the common data for picture primitives simplifies the
code and makes it impossible to accidentally access an invalid
bounding rect for picture primitives.

Differential Revision: https://phabricator.services.mozilla.com/D70295

--HG--
extra : moz-landing-system : lando