Bug 1665805 - Fix native compositor accuracy issues due to large clip rect. r=nical
If a native compositor surface ends up with no effective clip rect,
the clip rect supplied via the compositor trait to CoreAnimation
may be very large.

This large clip rect can cause accuracy issues inside CoreAnimation
resulting in layers that are drawn incorrectly.

To work around this, clamp the surface clip rect to the current
allocated area of the tiles for a given compositor surface.

Differential Revision: https://phabricator.services.mozilla.com/D90638