Bug 1688820 - force blend_pixels to inline more aggressively. r=jrmuizel
It appears that the slight refactoring of blend_pixels from the clip-mask
optimization in bug 1688104 caused clang-cl to inline blend_pixels more
pessimistically than it was doing before. Since this is an extremely hot
function that all alpha-pass rendering relies on, we generally just want
to inline this regardless of the slight increase in compile size since
a failure to aggressively optimize this can have noticeable performance
impacts.

Differential Revision: https://phabricator.services.mozilla.com/D103033