Bug 1700151 - Detach subsurfaces on hide, r=gw
Detach subsurfaces when hiding by removing the subsurface object.
This allows us to reattach the subsurface to another parent surface,
i.e. another window, and/or recycle the subsurface within a surface pool.

Recycling subsurfaces is until now a quite rare thing in the Wayland
world. Thus adding it to the example-compositor allows to stress-
test Wayland compositors.

Depends on D110745

Differential Revision: https://phabricator.services.mozilla.com/D110863