Bug 1704694 - Add BatchFilter struct. r=gfx-reviewers,lsalzman
Previously, the filter used to select which batches a primitive
was added to was simply a rectangle in picture space. In future
it will include information about the z-slice a primitive is on.

To handle this, add a BatchFilter struct which we can expand in
future patches with more information. Also switch the PrimitiveVisibility
and DirtyRegion structs to use the new type.

Differential Revision: https://phabricator.services.mozilla.com/D111779