Bug 1710979 - Store scene clip instances in a separate array r=gfx-reviewers,lsalzman
Previously, these were stored as a SmallVec<2> within each
defined clip template. However, this leads to inefficient
memory usage for most cases.

This is particularly noticeable on youtube.com, which currently
defines a large number of clips.

With this patch, the allocated memory for clips in the display
list drops from ~820 kB to ~450 kB (we should also investigate
and reduce the number of clips defined on this page too).

Differential Revision: https://phabricator.services.mozilla.com/D115039