Bug 1715746 - Work around Adreno shader compilation crash. r=jrmuizel
On certain adreno devices we encountered frequent driver crashes
during glLinkProgram. The crash appears to be caused by a combination
of compiling a shader with some GLSL code the driver doesn't like,
compiling shaders concurrently in multiple threads (eg webrender's
renderer thread and Fenix's Android UI render thread), and said
threads having a certain stack address.

Specifically, the GLSL code the driver doesn't like is a vector
bitwise-and followed by a conversion to bvec, which occurs in many of
webrender's shaders.

While we have control over webrender's renderer thread, we are unable
to synchronize with or control the stack address of the UI render
thread. Fortunately, we can rewrite the GLSL code to use scalar
bitwise-ands, therefore avoiding the crash.

Differential Revision: https://phabricator.services.mozilla.com/D117427