Bug 1715935 - Add experimental low quality pinch-zoom mode. r=gfx-reviewers,kvark
Building on previous patches, this adds a preference to enable
a low quality pinch-zoom mode that reuses cached tiles where
possible.

It doesn't currently redraw at the higher quality scale at the
end of a pinch-zoom, and there are some sampling artifacts in some
cases. However, it should be enough to do some performance testing
on low and devices and see if it helps.

Differential Revision: https://phabricator.services.mozilla.com/D117497