Bug 1719232 - Simplify check for empty valid rect. r=jrmuizel
We know from testing that the device valid rect itself can sometimes
be empty which produces the symptom that the compositor valid rect
is empty after quantization to i32. So it is sufficient to just check
if that device rect is empty, rather than having to use get_surface_rect
to reproduce the math downstream that produces the compositor valid
rect.

Differential Revision: https://phabricator.services.mozilla.com/D121453