Bug 1742031 - Pt 1 - Reduce per-primitive work during visibility pass r=gfx-reviewers,nical
The per-primitive visibility calculations done during frame
building are (mostly) redundant. Since picture caches are always
enabled, and since we must determine what tile(s) a primitive is on
for dependency updates, that gives us all the information we need
for visibility (by determining at the _tile_ level which tiles are
dirty and/or visible).

This patch is a small step towards that, moving the main picture
cache update out of the update_prim_visibility functions. Follow
up patches will remove most of the per-primitive work done during
this pass.

Differential Revision: https://phabricator.services.mozilla.com/D131580