Bug 1745878 - Temporary fix for incorrect surface opaqueness test. r=gfx-reviewers,nical
Revert "Bug 1742742 - Use backdrop candidates for opaque picture detection r=gfx-reviewers,nical"

This reverts commit 6d918f342ae9dedaed4b411efe7998342f00fcf3.

The patch above can incorrectly determine a surface to be opaque
when it may not be, due to a child picture surface having a rect
that extends outside the parent estimate picture rect. The problem
is not the patch itself, but that it exposes an existing bug with
how bounding rects of child surfaces are handled.

I'm currently refactoring the visibility pass to fix how child
bounding rect propagation and inflation works. Because of this,
it's simpler to revert this change for now, and re-land once the
rest of the visibility pass work is complete.

Differential Revision: https://phabricator.services.mozilla.com/D133971