Bug 1766056 - Fix YUV color matrix for SWGL slow and fast paths. r=jrmuizel
The brush_yuv_image slow path expected a YUV color matrix that applies scaling,
whereas we set up a matrix for SWGL that had no scaling to support the fast path.
To work around the fact that both paths have different scaling requirements, we
pass the matrix with scaling down to the fast path and undo the scaling later
when constructing SWGL's internal YUVMatrix from it. This allows both paths to
just use the same matrix and avoids the footgun.

Differential Revision: https://phabricator.services.mozilla.com/D144479