Bug 1766985 - Add bias to lookup table index calculation for component transfer filters. r=gfx-reviewers,nical
For table and discreet component transfer filters, we generate a
256-item lookup table and store it in the GPU cache. The shader then
calculates which index to look up based on the input texture, using
the following equation:

k = int(floor(colora[i]*255.0));

When the table was specified with 2 values (as is one of the items in
our filter-component-transfer.yaml wrench test), then as per the spec
we'd expect an input value of 0.5 to select the second value. However,
with the above equation, this would in fact select the first.

The only reason this test currently passes on most platforms is
because the 0.5 input color gets converted from a float to a
uint (rounded to 128) then and back to a float before being rendered
by webrender. This means the input value is actually 0.5019608 rather
than 0.5, which multiplied by 255 is 128.000004. On some Android
devices we have noticed that 0.5019608 in fact gets truncated to
0.50196, which multiplied by 255 is 127.9998, meaning the wrong table
entry gets selected even despite the conversion to and from uint.

While this fixed-size table implementation will likely always have
edge cases, we can fix this relatively obvious one by adding a bias of
0.5 to the index calculation, eg:

k = int(floor(colora[i]*255.0 + 0.5));

This means that the correct value gets rendered on all platforms
regardless of float precision, and even if the conversion to and from
uint were avoided.

Differential Revision: https://phabricator.services.mozilla.com/D145075