Bug 1792568 [Wayland] Force full redraw of RenderCompositorSWGL on Resume() r=lsalzman
In Bug 1645677 we disable rendering and free compositor resources on GTK unmap signal when GtkWindow is hidden.
That ensures we create EGLSurface again when GtkWindow is shown and we use recent wl_surface on Wayland.

We use Resume() compositor signal to force compositor to re-create EGLSurface over new wl_surface.

RenderCompositorSWGL uses different mechanism but underlying WindowSurfaceWaylandMB is also deleted on unmap
and we need to request full repaint on Resume() as new WindowSurfaceWaylandMB was created.

Differential Revision: https://phabricator.services.mozilla.com/D158322