Bug 1829592 - Suppress remote texture ready check when WebRenderAPI::FlushPendingWrTransactionEventsWithoutWait() is used r=gfx-reviewers,lsalzman
WebRenderAPI::FlushPendingWrTransactionEventsWithoutWait() does not wait remote texture ready. It should be OK, since it is called only in WebRenderBridgeParent::ClearResources(). But there were cases that hit the assert. The assert failure was false alarm.

The change suppresses assert of remote texture ready check in RemoteTextureMap::GetExternalImageIdOfRemoteTexture() if wait of the remote texture ready was skipped by the FlushPendingWrTransactionEventsWithoutWait().

Differential Revision: https://phabricator.services.mozilla.com/D176498