Bug 1849661 - Use new clip-mask render paths for picture primitives r=gfx-reviewers,nical
This patch removes use of the legacy clip-mask paths for picture
primitives. It makes use of render task sub-pass functionality
to apply the mask directly on top of the picture primitive when
the coordinate systems match (the common case). In the case of
complex transforms it renders them to a screen-space mask which
can be applied by the old clip-mask sampling code. In future,
we'll likely apply these as a sub-pass too.

Differential Revision: https://phabricator.services.mozilla.com/D186638