Bug 1604280 - Implement an APZ minimap for WebRender. r=dlrobertson,gw
The APZ minimap is a debugging tool for Firefox developers, which can
be enabled with the pref apz.minimap.enabled.

When enabled, it shows a minimap overlay for every scroll frame, with
rects indicating the layout and visual viewports and the displayport.

We previously had an implementation for Layers that was removed when
Layers was removed. This patch is restoring an equivalent feature,
implemented for WebRender.

Differential Revision: https://phabricator.services.mozilla.com/D190720