Bug 1883064 - Re-rasterize picture cache tiles when scale doubles or halves during low-quality pinch zoom. r=gw
When low-quality-pinch-zoom is enabled we avoid updating the raster
scale whilst a zoom is active, in order to avoid expensive
re-rasterizations. As a result, however, zooming out a very long
way (ie from fully zoomed in to fully zoomed out) causes a huge number
of picture cache tiles to be on screen at once. Allocating this number
of textures causes extremely bad jank and potentially even OOM
crashes.

Additionally, when zooming in a very long way content appears very
blurry.

To avoid this, this patch updates the raster scale during a
low-quality pinch zoom whenever the zoom has doubled or halved since
the raster scale was last updated. This avoids the blurriness as well
as limiting the number of tiles on screen at once. This comes at the
cost of the occasional expensive frame when everything is
re-rasterized, but is a worthwhile trade-off.

Differential Revision: https://phabricator.services.mozilla.com/D203287