Bug 1884791 - Avoid shader miscompilation on some Adreno drivers. r=gw
Webrender's glslopt-optimized shaders encounter a miscompilation on
some Adreno driver versions regarding fetching empty clip tasks. This
patch reshuffles the code in such a way as to avoid the
bug. Unfortunately the specific cause of the miscompilation remains
unknown, meaning we must take extra care not to regress it in the
future.

Differential Revision: https://phabricator.services.mozilla.com/D204864