Bug 1920058 - Add a debug pref to show picture borders. r=gfx-reviewers,lsalzman
This patch moves the logic to draw the tiled picture overlay out of the entirely too large take_context into a separate function, and add a separate option to draw only the borders of all pictures (non-tiled ones included) and tiles. This new overlay is enabled via the pref gfx.webrender.debug.picture-borders and is useful when trying to observe the breakdown of the page into pictures.
It does not quite do what I want it to: borders for non-tiled pictures seem to only appear on the frame they are drawn and aren't shown in subsequent frames unles the containing tile cache is invalidated. So there will probably be followup adjustments, but the functionality is already useful in its current state.

Differential Revision: https://phabricator.services.mozilla.com/D222959